requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
alembic>=1.10.0
psycopg2-binary
python-dotenv
passlib==1.7.4
bcrypt==4.0.1
python-jose
python-multipart